[YS-176] refactor: 유틸 함수 및 훅 파일 분리
